% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff.R
\name{ckn.create.fixed}
\alias{ckn.create.fixed}
\title{Fixed-X knockoffs}
\usage{
ckn.create.fixed(X, intercept = F, method = c("sdp", "equi"), randomize = F)
}
\arguments{
\item{X}{normalized n-by-p matrix of original variables.(\eqn{n \geq p}).}

\item{method}{either "equi" or "sdp" (default: "sdp").
This determines the method that will be used to minimize the correlation between the original variables and the knockoffs.}

\item{randomize}{whether the knockoffs are constructed deterministically or randomized (default: F).}
}
\value{
An object of class "knockoff.variables". This is a list
containing at least the following components:
\item{X}{n-by-p matrix of original variables (possibly augmented or transformed).}
\item{Xk}{n-by-p matrix of knockoff variables.}
}
\description{
Creates fixed-X knockoff variables.
}
\details{
Fixed-X knockoffs assume a homoscedastic linear regression model for \eqn{Y|X}. Moreover, they only guarantee
FDR control when used in combination with statistics satisfying the "sufficiency" property.
In particular, the default statistics based on the cross-validated lasso does not satisfy this
property and should not be used with fixed-X knockoffs.
}
\examples{

}
\references{
Barber and Candes,
Controlling the false discovery rate via knockoffs.
Ann. Statist. 43 (2015), no. 5, 2055--2085.
}
