% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff.R
\name{create.fixed.MRC}
\alias{create.fixed.MRC}
\title{Title}
\usage{
create.fixed.MRC(
  X,
  method = c("mvr", "maxent", "mmi", "sdp", "equicorrelated", "ci"),
  solver = NULL,
  knockpy = NULL,
  num_processes = 1
)
}
\arguments{
\item{num_processes}{}
}
\value{
a list of (X, X_kn)
}
\description{
Title
}
