% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff.R
\name{create_equicorrelated}
\alias{create_equicorrelated}
\title{Create equicorrelated fixed-X knockoffs.}
\usage{
create_equicorrelated(X, intercept, randomize)
}
\description{
Create equicorrelated fixed-X knockoffs.
}
\keyword{internal}
