% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cknockoff.R
\name{print.cknockoff.result}
\alias{print.cknockoff.result}
\title{Print results of cknockoff}
\usage{
\method{print}{cknockoff.result}(x, ...)
}
\arguments{
\item{x}{the output of a call to cknockoff}

\item{...}{unused}
}
\description{
Prints the list of variables selected by cknockoff and the corresponding function call.
}
