% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_X}
\alias{process_X}
\title{Compute and store matrices needed by cknockoff() from on X}
\usage{
process_X(X, knockoffs = ckn.create.fixed, intercept = T)
}
\arguments{
\item{X}{X n-by-p matrix of original variables.}

\item{knockoffs}{either knockoff matrix of X or a knockoffs function that can
generate it.}

\item{intercept}{Will intercept be fitted (default=TRUE) or set to zero
(FALSE).
If the knockoff matrix is supplied, both X and knockoffs should be properly
normalized, e.g. using the returned X and Xk of the ckn.create.fixed function.
By default, ckn.create.fixed is used.}
}
\value{
An object of class "cknockoff.X.pack". This object is a list containing
many matrices like the knockoff matrix and a basis of the linear space spanned by X, etc.
Users don't need to work on the object themselves but pass it to the "X.pack"
parameter in cknockoff() if needed.
}
\description{
Compute and store matrices needed by cknockoff() from on X
}
\examples{
p <- 100; n <- 300; k <- 15
X <- matrix(rnorm(n*p), n)
nonzero <- sample(p, k)
beta <- 2.5 * (1:p \%in\% nonzero)
y <- X \%*\% beta + rnorm(n)
print(which(1:p \%in\% nonzero))

X.pack <- process_X(X, knockoffs = ckn.create.fixed)

result <- cknockoff(X, y,
                    alpha = 0.05,
                    n_cores = 1,
                    X.pack = X.pack)
print(result$selected)


}
