% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_stats.R
\name{stat.glmnet_lambdasmax_lm}
\alias{stat.glmnet_lambdasmax_lm}
\title{Importance statistics based on Lasso penalty and Lagrange Multipliers}
\usage{
stat.glmnet_lambdasmax_lm(X, X_k, y, sigma_tilde = NULL)
}
\arguments{
\item{X}{n-by-p matrix of original variables.}

\item{X_k}{n-by-p matrix of knockoff variables.}

\item{y}{y vector of length n, containing the response variables.}

\item{sigma_tilde}{An estimator of the noise level independent of
\eqn{[X, X_k]^T y}.
By default, it is set to be NULL and will be computed inside the function
based on X, X_k, and y.}
}
\value{
A vector of statistics \eqn{W} of length p.
}
\description{
Importance statistics based on Lasso penalty and Lagrange Multipliers
}
\details{

}
\seealso{
Other statistics: 
\code{\link{stat.glmnet_coefdiff_lm}()}
}
\concept{statistics}
