% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff.R
\name{kn.select}
\alias{kn.select}
\title{Title}
\usage{
kn.select(kn_statistics, alpha, selective = T, early_stop = F)
}
\arguments{
\item{kn_statistics}{Knockoff W-statistics}

\item{alpha}{nominated FDR level}

\item{selective}{T: use selective SeqStep, F: use SeqStep}

\item{early_stop}{T/F}
}
\value{
list(selected, estimated fdp, |W_(k hat)|)
}
\description{
Title
}
